+++ .eslintrc.js
@@ -19,3 +19,4 @@ module.exports = {
 		"no-unused-vars": ["off"],
 	},
 };
+