+++ CHANGELOG.md
@@ -269,3 +269,9 @@
  - Added fbLink to "Flag" if fbLink is invalid in /profile
  - Added MissingData field to /profile/update
  - All Response Body should have a new token field which contains token_type and access_token (if role has changed)
+
+   ### 2023-02-28 22:12 (Maxx) ###
+ **Changed**
+ - Reformatted spec.yaml (still have a few to go, only finished /signup and /login)
+ **Planned**
+ - Remake data stored in AccessToken